change name to 1.0 and rebase
