+++ instrumentation/restlet/restlet-1.0/javaagent/src/test/groovy/io/opentelemetry/javaagent/instrumentation/restlet/v1_0/RestletServerTest.groovy
@@ -1,0 +1,23 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.restlet.v1_0
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import io.opentelemetry.instrumentation.restlet.v1_0.AbstractRestletServerTest
+import io.opentelemetry.instrumentation.test.AgentTestTrait
+import org.restlet.Restlet
+
+class RestletServerTest extends AbstractRestletServerTest implements AgentTestTrait {
+
+  @Override
+  Restlet wrapRestlet(Restlet restlet, String path){
+    return restlet
+  }
+
+}
