--- instrumentation/restlet/restlet-1.1/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/restlet/v1_1/RestletInstrumentationModule.java
@@ -1,25 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.restlet.v1_1;
-
-import com.google.auto.service.AutoService;
-import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import java.util.Arrays;
-import java.util.List;
-
-@AutoService(InstrumentationModule.class)
-public class RestletInstrumentationModule extends InstrumentationModule {
-
-  public RestletInstrumentationModule() {
-    super("restlet", "restlet-1.1");
-  }
-
-  @Override
-  public List<TypeInstrumentation> typeInstrumentations() {
-    return Arrays.asList(new ServerInstrumentation(), new RouteInstrumentation());
-  }
-}
