--- instrumentation/restlet/restlet-1.1/library/src/test/groovy/RestletServerTest.groovy
@@ -1,24 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import io.opentelemetry.instrumentation.restlet.v1_1.AbstractRestletServerTest
-import io.opentelemetry.instrumentation.restlet.v1_1.TracingFilter
-import io.opentelemetry.instrumentation.test.LibraryTestTrait
-import org.restlet.Restlet
-
-class RestletServerTest extends AbstractRestletServerTest implements LibraryTestTrait {
-
-  @Override
-  Restlet wrapRestlet(Restlet restlet, String path){
-    def filter = new TracingFilter(path)
-    filter.setNext(restlet)
-    return filter
-  }
-
-  @Override
-  boolean testException() {
-    false //Filter's afterHandle does not execute if exception was thrown in the next restlet
-  }
-}
